// blue-cat-studio/src/contracts/enums/meta/effect/attribute-type.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export enum AttributeType {
  AttackDamage = "AttackDamage",
  FirePower = "FirePower",
  IcePower = "IcePower",
  EarthPower = "EarthPower",
  DarkPower = "DarkPower",
  LightPower = "LightPower",
  DamageResistance = "DamageResistance",
  FireResistance = "FireResistance",
  IceResistance = "IceResistance",
  EarthResistance = "EarthResistance",
  DarkResistance = "DarkResistance",
  LightResistance = "LightResistance",
  MoveSpeed = "MoveSpeed",
  AttackSpeed = "AttackSpeed",
  AttackRange = "AttackRange",
  Lucky = "Lucky",
  Health = "Health",
  Stamina = "Stamina",
  Energy = "Energy",
}